requirements.txt
# Core data science and ML libraries
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.11.0
lightgbm>=4.0.0
xgboost>=2.0.0

# Data visualization
matplotlib>=3.7.0
plotly>=5.15.0
seaborn>=0.12.0

# Web framework
streamlit>=1.28.0

# Financial data sources
yfinance>=0.2.0
fredapi>=0.5.0
requests>=2.31.0
python-dotenv>=1.0.0
alpaca-py>=0.13.0

# LLM / NLP
openai>=1.40.0

# Trading calendar (market holidays and trading days)
pandas-market-calendars>=4.3.0
# Alternative: exchange-calendars>=4.2.0  # Uncomment if you prefer exchange-calendars

# Professional backtesting
bt>=1.1.0

# Configuration and validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database and storage
sqlalchemy>=2.0.0

# Utilities
pathlib>=1.0.1
typing-extensions>=4.8.0
lxml>=4.9.0  # For pandas.read_html
finnhub>=2.4.19

# Development and testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.1.0
mypy>=1.7.0

# Optional: Deep learning capabilities
torch>=2.0.0

# Documentation
sphinx>=7.2.0

# Packaging
setuptools>=68.0.0
wheel>=0.41.0
